Made a Scorebook Svelte component.
